Move stats to Model Package
