+++ src/main/java/seedu/weme/logic/Logic.java
@@ -14,8 +14,8 @@ import seedu.weme.logic.prompter.prompt.CommandPrompt;
 import seedu.weme.model.ModelContext;
 import seedu.weme.model.ReadOnlyWeme;
 import seedu.weme.model.meme.Meme;
+import seedu.weme.model.statistics.Stats;
 import seedu.weme.model.template.Template;
-import seedu.weme.statistics.Stats;
 
 /**
  * API of the Logic component
