+++ src/main/java/seedu/weme/model/ReadOnlyWeme.java
@@ -5,10 +5,10 @@ import java.util.List;
 import javafx.collections.ObservableList;
 
 import seedu.weme.model.meme.Meme;
+import seedu.weme.model.statistics.Stats;
+import seedu.weme.model.statistics.TagWithCount;
 import seedu.weme.model.tag.Tag;
 import seedu.weme.model.template.Template;
-import seedu.weme.statistics.Stats;
-import seedu.weme.statistics.TagWithCount;
 
 /**
  * Unmodifiable view of a Weme
@@ -48,7 +48,7 @@ public interface ReadOnlyWeme {
      * Returns the count of a tag in the memeList.
      * Returns -1 if the tag is not present in the memeList.
      */
-    public int getCountOfTag(Tag tag);
+    int getCountOfTag(Tag tag);
 
     /**
      * Returns a list of tags with their counts.
