+++ src/main/java/seedu/weme/model/statistics/Stats.java
@@ -1,4 +1,4 @@
-package seedu.weme.statistics;
+package seedu.weme.model.statistics;
 
 import java.util.List;
 import java.util.Map;
